refactor: 레이싱 게임에서 입출력 기능 분리
