+++ src/main/java/racingcar/RacingGameApplication.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.controller.RacingGameController;
+
 public class RacingGameApplication {
     public static void main(String[] args) {
         RacingGameController racingGameController = new RacingGameController();
